comments/serializers.py
from rest_framework import serializers

from .models import Comment


class CommentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Comment
        fields = ['author', 'pub_date', 'text',
        'nesting_level', 'parent_comment']
